% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BevHolt.R
\name{BevHolt}
\alias{BevHolt}
\title{Beverton-Holt function}
\usage{
BevHolt(x)
}
\arguments{
\item{x}{string of components for BH function}
}
\value{
NLL for fitting BH
}
\description{
Beverton-Holt function
}

