% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brw.R
\name{brw}
\alias{brw}
\title{biased random walk function}
\usage{
brw(
  n = 1,
  data,
  xy = NULL,
  coa = NULL,
  dir = NULL,
  buffer = NULL,
  rho,
  a,
  b,
  shelf,
  beta
)
}
\description{
utility function not to be called by user
}
