% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBHrec.R
\name{calcBHrec}
\alias{calcBHrec}
\title{calculate recruits per beverton-holt}
\usage{
calcBHrec(SRouts, spbioIn)
}
\arguments{
\item{SRouts}{spawner recruit parameters}

\item{spbioIn}{spawning biomass}
}
\value{
predicted recruits
}
\description{
calculate recruits per beverton-holt
}

