% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcREF.R
\name{calcREF}
\alias{calcREF}
\title{\code{calcREF} calculates reference points}
\usage{
calcREF(SRouts)
}
\arguments{
\item{SRouts}{presumably spawner per recruit outs}
}
\value{
unclear
}
\description{
\code{calcREF} calculates reference points
}

