% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRickRec.R
\name{calcRICKrec}
\alias{calcRICKrec}
\title{Calculate ricker recruits}
\usage{
calcRICKrec(SRouts, spbioIn)
}
\arguments{
\item{SRouts}{parameters of ricker recruitment}

\item{spbioIn}{spawning biomass}
}
\value{
predicted ricker recruits
}
\description{
Calculate ricker recruits
}

