% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanInput.R
\name{CleanInput}
\alias{CleanInput}
\title{\code{CleanInput} extends parameters for the right amount of time}
\usage{
CleanInput(input, SimYear)
}
\arguments{
\item{input}{the thing to be extended}

\item{SimYear}{the number of years to extend}
}
\value{
the extended object
}
\description{
\code{CleanInput} extends parameters for the right amount of time
}

