% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_route.R
\name{find_route}
\alias{find_route}
\title{find_route}
\usage{
find_route(
  data,
  mpar,
  coas = 12,
  npts = 40,
  land.buff = 2,
  buffer = 5,
  centroids = FALSE,
  keep.coas = NULL
)
}
\arguments{
\item{data}{\itemize{
\item a list of required data. If missing then simulation runs on a
Cartesian grid (a featureless environment).
}}

\item{mpar}{\itemize{
\item simulation control parameters supplied as a list using \code{sim_par()}
}}

\item{coas}{\itemize{
\item number of Centres of Attraction to calculate
}}

\item{npts}{\itemize{
\item number of route-finding points to use. A larger number of points
can produce smoother routes that do not jump across land, a smaller number of
points will produce a less complex route, faster.
}}

\item{land.buff}{\itemize{
\item distance (km) to buffer around land. A small buffer
(e.g., 1 - 2 km) can reduce coastline complexity and make route-finding simpler.
}}

\item{buffer}{\itemize{
\item \code{pathroutr} argument that sets the typical distance from
coastline of the visibility graph.
}}

\item{centroids}{\itemize{
\item \code{pathroutr} argument that turns on/off (default: FALSE)
centroids in calculating the visibility graph
}}

\item{keep.coas}{\itemize{
\item index of calculated CoA's to use in the simulation. Default
is NULL (use all calculated CoA's).
}}
}
\value{
a list including a matrix of calculated CoA's and a \code{ggplot2} object
visualizing the CoA's. The visualization is useful for determining whether to
use all the CoA's or subset them, and for helping to diagnose route-finding
failures.
}
\description{
check for barriers between start and end (CoA) points & find route
around barriers via intermediate CoA's to be used by \code{sim_fish()}. Can be
called prior to \code{sim_fish()} or within \code{sim_fish()}.
}
