% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_land_sf.R
\name{gen_land_sf}
\alias{gen_land_sf}
\title{utility function - generate sf land}
\usage{
gen_land_sf(ext, dist = 0)
}
\arguments{
\item{ext}{longitude and latitude extents as a 4-element vector:
(xmin, ymin, xmax, ymax)}

\item{dist}{buffer distance (km) to add to land. This can be used when
generating the gradient rasters to keep tracks a specified distance from land.
The default is 0 km, no buffer.}
}
\value{
an sf data.frame defining the simulation environment (i.e., ocean with
land barriers).
}
\description{
generates an sf data.frame of land from user-specified extents
}
\keyword{internal}
