% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_env.R
\name{generate_env}
\alias{generate_env}
\title{create environmental raster}
\usage{
generate_env(
  ext = NULL,
  res = c(0.05, 0.05),
  prj = "+proj=merc +datum=WGS84 +units=km"
)
}
\arguments{
\item{ext}{longitude and latitude extents as a 4-element vector:
(xmin, ymin, xmax, ymax)}

\item{res}{resolution of raster in degrees long and lat. Smaller values
yield higher resolution but longer computation times.}

\item{prj}{a crs string or EPSG defining the projection for the raster with
units in km. The default projection is a Mercator grid:
\code{prj = "+prj=merc +datum=WGS84 +units=km"}}

\item{...}{additional arguments to be ignored}
}
\value{
a raster defining the simulation environment
}
\description{
create environmental raster to constraint simulated tracks
}
\examples{
x <- generate_env(ext = c(-70,43,-52,53), res = c(0.04,0.04))

raster::plot(x, zlim = c(0,1))

}
\author{
Ian Jonsen \email{ian.jonsen@mq.edu.au}
}
