% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_grad.R
\name{generate_grad}
\alias{generate_grad}
\title{create gradient rasters for potential function}
\usage{
generate_grad(x)
}
\arguments{
\item{x}{input environmental raster}
}
\value{
two gradient SpatRasters defining gradients in distance from water
}
\description{
create gradient rasters for potential function that constrains
simulated to stay off land. Uses the `terra` package instead of `raster` for
enhanced computational speed
}
\keyword{internal}
