% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HarvestControlRules.R
\name{HarvestControlRule}
\alias{HarvestControlRule}
\title{\code{HarvestControlRule} runs a harvest control rule}
\usage{
HarvestControlRule(FMSY, BMSY, ExploitBio, SpawnBio, alpha, beta,
  HarvestControl, ConstantCatch = 0, ConstantF = 0)
}
\arguments{
\item{FMSY}{Fmsy}

\item{BMSY}{Bmsy}

\item{ExploitBio}{the exploitable biomass}

\item{SpawnBio}{the spawning biomass}

\item{alpha}{alpha something}

\item{beta}{beta something}

\item{HarvestControl}{form of harvest control rule}

\item{ConstantCatch}{constant catch}

\item{ConstantF}{constant F}
}
\value{
a TAC
}
\description{
\code{HarvestControlRule} runs a harvest control rule
}

