% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialN.R
\name{initialN}
\alias{initialN}
\title{Calculate initial numbers}
\usage{
initialN(Rzero, NatM, inAge)
}
\arguments{
\item{Rzero}{recruits at EQ}

\item{NatM}{natural mortality}

\item{inAge}{no idea}
}
\value{
initial numbers
}
\description{
Calculate initial numbers
}

