% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\title{map}
\usage{
map(x, env = NULL, by.id = FALSE, ...)
}
\arguments{
\item{x}{a simfish simulation object}

\item{env}{...}

\item{by.id}{logical (default: FALSE); if x is a multi-track object then colour tracks by id}

\item{...}{additional arguments to be ignored}
}
\value{
a ggplot object
}
\description{
a map of single or multiple simulated tracks
}
\author{
Ian Jonsen \email{ian.jonsen@mq.edu.au}
}
