% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_kernel.R
\name{move_kernel}
\alias{move_kernel}
\title{random walk movement kernel for Miramichi smolts}
\usage{
move_kernel(data, xy = NULL, mpar, i, s, ts, d2l, dir2l)
}
\description{
utility function not to be called by user
}
