% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.simfish.R
\name{plot.simfish}
\alias{plot.simfish}
\title{plot}
\usage{
\method{plot}{simfish}(x, by.id = FALSE, ...)
}
\arguments{
\item{x}{a simfish simulation object}

\item{by.id}{logical (default: FALSE); if x is a multi-track object then colour tracks by id}

\item{...}{additional arguments to be ignored}
}
\value{
a ggplot object
}
\description{
a simple visualization of single or multiple simulated tracks
}
\examples{
my.par <- sim_par(N = 1440, time.step = 5, start = c(0, 0), coa = c(0,30))

z <- sim_fish(id = 1, mpar = my.par)
plot(z)

}
\author{
Ian Jonsen \email{ian.jonsen@mq.edu.au}
}
