% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recruitment.R
\name{Recruitment}
\alias{Recruitment}
\title{Calculate recruitment}
\usage{
Recruitment(EggsIN, steepnessIN, RzeroIN, RecErrIN, recType, NatMin, vulnIN,
  matureIN, weightIN, LenAtAgeIN, MaxAge)
}
\arguments{
\item{EggsIN}{eggs}

\item{steepnessIN}{steepness}

\item{RzeroIN}{eq recruitment}

\item{RecErrIN}{recruitment error}

\item{recType}{form of recruitment}

\item{NatMin}{natural mortality}

\item{vulnIN}{fishing vulnerability}

\item{matureIN}{maturity at age}

\item{weightIN}{weight at age}

\item{LenAtAgeIN}{length at age}

\item{MaxAge}{maximim age}
}
\value{
returns the recruits
}
\description{
\code{Recruitment} returns the recruits for a given functional form and SSB
}

