% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAnalysis.R
\name{runAnalysis}
\alias{runAnalysis}
\title{Run the fish density simulation}
\usage{
runAnalysis(Mcmc_prediction = TRUE, Abundance_or_catch = "Catch",
  Mcmc_sample = 1, Mcmc_seed = 200, Which_species, Which_region,
  Years = NULL, Env_covariate = c("SURFACE_TEMPERATURE",
  "GEAR_TEMPERATURE", "BOTTOM_DEPTH"), Transform_covariate = c("ID",
  "ID", "LOG"), Include_year_effect = FALSE, Include_2_order = TRUE,
  Standardize_variable = FALSE, Spacetime = "AR",
  YEARS_pred = 1982:2016, Overwrite = FALSE)
}
\arguments{
\item{year}{Either a single year or 2 element vector with maximum and minimum year to retrieve.}

\item{survey}{String for intended survey (EBS_SHELF, EBS_SLOPE, NBS_SHELF, GOA, AI).}
}
\value{
The output is a list of years and densities. The output must be re-joined to the Predict_data
object for spatial analysis.
  \item{YEARS_USE}{Vector of available years}
  \item{SDM_result}{Matrix of densities per year, rows for a given year correspond to Predict_data rows.}
}
\description{
Executes the simulation for a given region and species.
}
\examples{
## TBD

}
