% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBPRfunc.R
\name{SBPRfunc}
\alias{SBPRfunc}
\title{Calc Spawning biomass per recruit}
\usage{
SBPRfunc(F, Rzero, NatM, vuln, mature, weight, LenAtAge, MaxAge)
}
\arguments{
\item{Rzero}{recruitment at eq}

\item{NatM}{natural mortality}

\item{vuln}{vulnerability to fishing}

\item{mature}{maturity at age}

\item{weight}{weight at age}

\item{LenAtAge}{length at age}
}
\description{
\code{SBPRfunc} calculate spaning biomass per recruit based on
parameters
}

