% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_par.R
\name{sim_par}
\alias{sim_par}
\title{Define the simulation parameters used by \code{sim_fish}}
\usage{
sim_par(...)
}
\arguments{
\item{...}{simulation control parameters}
}
\value{
Returns a list of simulation parameters to be used by \code{sim_fish}.
See example, below
}
\description{
Define simulation parameters
}
\examples{
## A minimal example - simulation with no environment
my.par <- sim_par(N = 1440, time.step = 5, start = c(0, 0), coa = c(0,30))

out <- sim_fish(id = 1, mpar = my.par)
plot(out)


}
