% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_setup.r
\name{sim_setup}
\alias{sim_setup}
\title{Pre-simulation setup}
\usage{
sim_setup(config = config)
}
\arguments{
\item{config}{- path to config.R script containing file.paths for required & optional environmental layers (see Details)}
}
\description{
load required rasters, receiver locations
}
\author{
Ian Jonsen \email{ian.jonsen@mq.edu.au}
}
