% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_smolt.R
\name{sim_smolt}
\alias{sim_smolt}
\title{simulate smolt migration}
\usage{
sim_smolt(id = 1, data = NULL, mpar = sim_par(), pb = TRUE)
}
\arguments{
\item{id}{- identifier for simulation run (individual animal)}

\item{data}{- a list of required data from \code{presim}}

\item{mpar}{- simulation control parameters supplied as a list, see details}

\item{pb}{- use progress bar (logical)}
}
\description{
simulates smolt migration
}
\author{
Ian Jonsen \email{ian.jonsen@mq.edu.au}
}
