% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simf.R
\name{simf}
\alias{simf}
\title{simf}
\usage{
simf(data, mpar, s, xy, coas = FALSE)
}
\arguments{
\item{data}{- a list of required data. If missing then simulation runs on a
Cartesian grid (a featureless environment).}

\item{mpar}{- simulation control parameters supplied as a list using `sim_par()`
See `?sim_par` for details on the simulation parameters.}
}
\description{
simulates fish tracks in featureless or semi-realistic
environments. Workhorse fn called by front-end `sim_fish()`.
}
\author{
Ian Jonsen \email{ian.jonsen@mq.edu.au}
}
\keyword{internal}
