% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfish-package.R
\docType{package}
\name{simfish-package}
\alias{simfish-package}
\alias{simfish}
\title{\pkg{simfish}}
\description{
simulate fish tracks & acoustic detections
}
\seealso{
Useful Links:
* [https://github.com/ianjonsen/simfish/](https://github.com/ianjonsen/simfish/)
* Report bugs/issues at [https://github.com/ianjonsen/simfish/issues/](https://github.com/ianjonsen/simfish/issues/)
}
\author{
**Maintainer:** Ian Jonsen <ian.jonsen@mq.edu.au> [ORCID](https://orcid.org/0000-0001-5423-6076)
}
\keyword{simfish}
