% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StockRecCheck.R
\name{StockRecCheck}
\alias{StockRecCheck}
\title{StockRecCheck checks stock recruitment}
\usage{
StockRecCheck(rec, spbio, plotSR = F)
}
\arguments{
\item{rec}{recruitment}

\item{spbio}{spawning biomass}

\item{plotSR}{plot spawners per recruit}
}
\value{
list of checks
}
\description{
StockRecCheck checks stock recruitment
}

