% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_env.R
\name{generate_env}
\alias{generate_env}
\title{create environmental raster}
\usage{
generate_env(
  ext = NULL,
  res = c(0.05, 0.05),
  prj = "+proj=merc +datum=WGS84 +units=km",
  grad = FALSE,
  dist = 0
)
}
\arguments{
\item{ext}{longitude and latitude extents as a 4-element vector:
(xmin, ymin, xmax, ymax)}

\item{res}{resolution of raster in degrees long and lat. Smaller values
yield higher resolution but longer computation times.}

\item{prj}{a crs string or EPSG defining the projection for the raster with
units in km. The default projection is a Mercator grid:
\code{prj = "+prj=merc +datum=WGS84 +units=km"}.}

\item{grad}{(logical) generate gradient rasters.
The gradient rasters (gradient of distance from water in x and y directions)
are required by the potential function used to simulate tracks.}

\item{dist}{buffer distance (km) to add to land. This can be used when
generating the gradient rasters to keep tracks a specified distance from land.
The default is 0 km, no buffer.}
}
\value{
a raster defining the simulation environment (i.e., ocean with
land barriers). Optionally, the gradient rasters, required for the simulation
potential function, can also be generated. Only one of the environment raster
or the gradient rasters can be returned.
}
\description{
create environmental raster to constraint simulated tracks
}
\examples{
x <- generate_env(ext = c(-70,43,-52,53), res = c(0.04,0.04))

raster::plot(x)

}
\author{
Ian Jonsen \email{ian.jonsen@mq.edu.au}
}
