% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_grad.R
\name{generate_grad}
\alias{generate_grad}
\title{create gradient rasters for potential function}
\usage{
generate_grad(x)
}
\arguments{
\item{x}{input environmental raster}
}
\value{
two gradient rasters defining gradients in distance from water
}
\description{
create gradient rasters for potential function that constrains
simulated to stay off land
}
\examples{
x <- generate_env(ext = c(-70,43,-52,53), res = c(0.04,0.04))
g <- generate_grad(x)

raster::plot(g)

}
\author{
Ian Jonsen \email{ian.jonsen@mq.edu.au}
}
