% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\title{map}
\usage{
map(
  x,
  env = NULL,
  by.id = FALSE,
  coas = FALSE,
  term.pts = TRUE,
  zoom = FALSE,
  ...
)
}
\arguments{
\item{x}{a simfish simulation object}

\item{env}{...}

\item{by.id}{logical (default: FALSE); if x is a multi-track object then
colour tracks by id}

\item{coas}{logical (default: FALSE); should CoA's be displayed on map}

\item{term.pts}{logical (default: TRUE); should terminal (start, end) points
be displayed on map}

\item{zoom}{logical (default: FALSE); should map be zoomed to extent of
simulated track, otherwise map uses extents from the land raster in \code{env}}

\item{...}{graphical parameters. Arguments that can be passed to
\code{ggplot2::coord_sf()}, such as \code{xlim = c(20,40)}.}
}
\value{
a ggplot object
}
\description{
a map of single or multiple simulated tracks
}
\author{
Ian Jonsen \email{ian.jonsen@mq.edu.au}
}
