% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_detect.r
\name{sim_detect}
\alias{sim_detect}
\title{simulate acoustic transmissions & detection, using \code{simulate} & \code{sim_setup} output}
\usage{
sim_detect(s, data, delay = c(50, 130), burst = 5, noise = 1)
}
\arguments{
\item{s}{- a deadsmolt class list containing output from sim_setup and sim_smolt/sim_kelt}

\item{delay}{- min & max time intervals (s) between transmissions}

\item{burst}{- duration of each transmission (s)}

\item{noise}{- range 0 - 1; simulate effect of noisy environment. Reduces detection prob w dist
by specified proportion; default = 1, no reduction}
}
\description{
simulates transmissions & detections along simulated track segments within a defined range of acoustic array(s)
}
\author{
Ian Jonsen \email{ian.jonsen@mq.edu.au}
}
