% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_fish.R
\name{sim_fish}
\alias{sim_fish}
\title{simulate a fish track}
\usage{
sim_fish(id = 1, data = NULL, mpar = sim_par(), pb = TRUE)
}
\arguments{
\item{id}{- identifier for simulation run (individual fish)}

\item{data}{- a list of required data}

\item{mpar}{- simulation control parameters supplied as a list using `sim_par()`}

\item{pb}{- use progress bar (logical)}
}
\description{
simulates fish tracks
}
\author{
Ian Jonsen \email{ian.jonsen@mq.edu.au}
}
