% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_fish.R
\name{sim_fish}
\alias{sim_fish}
\title{simulate a fish track}
\usage{
sim_fish(id = 1, data = NULL, mpar = sim_par(), pb = TRUE)
}
\arguments{
\item{id}{- identifier for simulation run (individual fish)}

\item{data}{- a list of required data. If missing then simulation runs on a
Cartesian grid (a featureless environment).}

\item{mpar}{- simulation control parameters supplied as a list using `sim_par()`
See `?sim_par` for details on the simulation parameters.}

\item{pb}{- use progress bar (logical)}
}
\description{
simulates fish tracks in featureless or semi-realistic
environments. Semi-realistic environments are user-defined rasters of water
bounded by land, which constrain the simulated tracks. User-supplied acoustic
receiver locations can be used to simulate acoustic detections based on a
user-defined detection probability with distance. Utility functions are
provided to generate water-land rasters (`generate_land()`), required
gradient rasters (`generate_grad()`), and to convert acoustic tag
detection range expectations into logistic regression parameters (`calc_pdrf()`).
}
\examples{
## A minimal example - simulation with no environment
my.par <- sim_par(N = 1440, time.step = 5, start = c(0, 0), coa = c(0,30))

out <- sim_fish(id = 1, mpar = my.par, pb = FALSE)

plot(out)

## Simulate in a semi-realistic environment
land <- generate_env(ext = c(-70,43,-52,53), res = c(0.05,0.05))
grad <- generate_grad(land)
x <- list(land = land, grad = grad)

my.par <- sim_par(N=400, time.step=60*6, start = c(-7260, 5930),
coa = c(-6300,6680), nu = 0.6, rho = 0.7)
out <- sim_fish(id = 1, data = x, mpar = my.par, pb = FALSE)

map(out, env = x)
}
\author{
Ian Jonsen \email{ian.jonsen@mq.edu.au}
}
