% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_par.R
\name{sim_par}
\alias{sim_par}
\title{Control Values for \code{simulate}.}
\usage{
sim_par(...)
}
\arguments{
\item{...}{simulation control parameters}
}
\value{
...
}
\description{
\code{sim_par} defines the simulation parameters & control scenarios used by \code{simulate}.
}
\details{
The movement process used predominantly in the simulation is
selected by the \code{move} argument.  Additional
parameters include: \code{temp} is movement temperature-dependent; \code{advect} do ocean
currents influence movement; \code{growth} do smolts grow in temp-dependent fashion;
\code{start.date};
\code{start} (location); \code{coa} centre of attraction (can be NULL);
\code{mdir} directional bias (a 2-element vector);
\code{rho} concentration of step directions (for wrapped-Cauchy, a 2-element vector); ...
}
