% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_setup.r
\name{sim_setup}
\alias{sim_setup}
\title{Pre-simulation setup}
\usage{
sim_setup(config = config, deadsmolt = FALSE, esrf = TRUE)
}
\arguments{
\item{config}{- path to config.R script containing file.paths for required & optional environmental layers (see Details)}

\item{deadsmolt}{- logical; turns on Miramichi & SoBI receiver locations if TRUE}

\item{esrf}{- logical; turns off receiver & ESRF polygons if FALSE}
}
\description{
load required rasters, receiver locations
}
\author{
Ian Jonsen \email{ian.jonsen@mq.edu.au}
}
